% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envDataLoad.R
\name{envDataLoad}
\alias{envDataLoad}
\title{Loading the data previously read in and saved by "envDataRead"}
\usage{
envDataLoad(filename = stop("'filename' must be given!"), dir.in = ".")
}
\arguments{
\item{filename}{The base of the filenames; i.e. if the data is saved in
"my_data_env.ffData" and "my_data_env.RData", then the 'filename'
should be "my_data".}

\item{dir.in}{The path to the directory where files were saved (defaults to
the current directory).}
}
\value{
A list of ffdf objects with the environmental data in numeric format.
}
\description{
This function loads the data from the saved .ffData and .RData files,
and prepares the data to subsequent analysis.
}
