% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envDataSubset.R
\name{envDataSubset}
\alias{envDataSubset}
\title{Subsetting the environmental data}
\usage{
envDataSubset(
  env.data = stop("You need to specify the data!", call. = FALSE),
  col.names,
  row.names,
  col.ids,
  row.ids,
  file.out = "my_data_part",
  dir.out = ".",
  overwrite = NULL
)
}
\arguments{
\item{env.data}{The dataset (class \code{env.data}) from which to extract data.}

\item{col.names}{Character vector with column names to extract (optional).}

\item{row.names}{Character vector with row names to extract (optional).}

\item{col.ids}{Numeric vector with column IDs (optional).}

\item{row.ids}{Numeric vector with row IDs (optional).}

\item{file.out}{The base for the output filename (default: "my_data_part").}

\item{dir.out}{The path to the directory where the output files will be saved.}

\item{overwrite}{Whether to overwrite the output files: if NULL (default),
will prompt the user to give answer; set to TRUE, will automatically
overwrite any existing files; and set to FALSE, will stop if the output
files exist.}
}
\value{
A matrix with the subset of data.
}
\description{
Extract a subset of environmental data by specifying rows and/or columns.
}
\section{Warning}{

The parameter \code{col.names} cannot be used jointly with \code{col.ids},
similarly for the rows!
}

