% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrows_ncolumns.R
\name{nrows}
\alias{nrows}
\alias{ncolumns}
\title{Number of rows/columns of env.data}
\usage{
nrows(obj)

ncolumns(obj)
}
\arguments{
\item{obj}{env.cont or env.cat data, read in by \link{envDataRead}.}
}
\value{
integer
}
\description{
\code{nrows} and \code{ncolumns} return the number of rows or columns present
in obj.
}
