% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.env.data.R
\name{print.env.cat}
\alias{print.env.cat}
\title{Info about the \code{env.cat} class}
\usage{
\method{print}{env.cat}(x, ...)
}
\arguments{
\item{x}{The \code{env.cat} object read in by \link{envDataRead}.}
}
\value{
\code{NULL}
}
\description{
Prints a short summary about the categorical environmental data.
}
