% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.env.data.R
\name{print.env.cont}
\alias{print.env.cont}
\title{Info about the \code{env.cont} class}
\usage{
\method{print}{env.cont}(x, ...)
}
\arguments{
\item{x}{The \code{env.cont} object read in by \link{envDataRead}.}

\item{...}{ignored}
}
\value{
\code{NULL}
}
\description{
Prints a short summary about the continuous environmental data.
}
