% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showRaw.R
\name{showRaw}
\alias{showRaw}
\title{Show raw environmental data}
\usage{
showRaw(
  data = stop("You need to specify the data!", call. = FALSE),
  rows = 1:5,
  columns = 1:5
)
}
\arguments{
\item{data}{Object of class \code{env.data}.}

\item{rows}{Vector containing either numbers of rows or row names to show.
By default, if no 'rows' or 'columns' argument is given, the function
returns a 5x5 matrix.}

\item{columns}{Vector containing either numbers of columns or column names
to show.}
}
\value{
Matrix (in ff format) with extracted data.
}
\description{
Extracting the raw measurements from the \code{env.data} object structure.
}
