% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.env.data.R
\name{summary.env.cont}
\alias{summary.env.cont}
\title{Info about the \code{env.cont} class}
\usage{
\method{summary}{env.cont}(object, short = TRUE, ...)
}
\arguments{
\item{object}{The \code{env.cont} object read in by \link{envDataRead}.}

\item{...}{ignored}
}
\value{
list with:
  \itemize{
    \item class - the full class name;
    \item nrow - number of rows;
    \item ncol - number of columns;
    \item rownames - character vector with row names;
    \item colnames - character vector with column names.
  }
}
\description{
Returns a short summary about the continuous environmental data.
}
