% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envDataCategorize.R
\name{envDataCategorize}
\alias{envDataCategorize}
\title{Creating categorical variables out of continuous data}
\usage{

  envDataCategorize(env.data = stop("You didn't provide the environmental data!",
  call. = FALSE), breaks, file.out = "env_data_cat", dir.out = ".",
  ncpu = 1, overwrite = NULL)
}
\arguments{
\item{env.data}{The environmental data, read in by \link{envDataRead}
function.}

\item{breaks}{Numerical vector indicating how to divide the continuous values
into categories (see Details).}

\item{file.out}{The core name of the files that will contain the categorized
data (character string); ready to load next time with \link{envDataLoad}
function; default: "env_data_cat".}

\item{dir.out}{The directory that will contain the saved data; defaults to
current working directory.}

\item{ncpu}{The number of CPU cores to use - this speeds up the process for
large datasets significantly. Default is 1 core, maximum is 1 less than
the total number of cores available on a current machine (even if the number
given by the user is more than that).}

\item{overwrite}{Whether to overwrite the output files: if NULL (default),
will prompt the user to give answer; set to TRUE, will automatically
overwrite any existing files; and set to FALSE, will stop if the output
files exist.}
}
\value{
A list of ff matrices, now containing the categorized data (factors).
}
\description{
This function prepares the environmental per-SNP data to be used in
  stratification when calling \link{haplinStrat}.
}
\section{Details}{

  When 'breaks' is one number, it defines the number of categories that the
  range of values will be divided into. This is good only for equally
  distributed data. Otherwise, the user is adviced to give 'breaks' as a
  vector of break points, similarly to when creating a histogram.
}

