% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envDataRead.R
\name{envDataRead}
\alias{envDataRead}
\title{Reading the environmental data from a file.}
\usage{
envDataRead(
  file.in = stop("'file.in' must be given!"),
  dir.in = ".",
  file.out = NULL,
  dir.out = ".",
  sep = ",",
  cont = TRUE,
  header = TRUE,
  rownames = TRUE,
  overwrite = NULL
)
}
\arguments{
\item{file.in}{The name of the delimited file with environmental data –
including the file extension (e.g., \code{"methyl_data.dat"}). The correct
separator must be specified via the \code{sep} argument.}

\item{dir.in}{The path to the directory where the \code{file.in} file
resides.}

\item{file.out}{The base name for the output files (see Details).}

\item{dir.out}{The path to the directory where the output files will be
saved.}

\item{sep}{The separator character that separates values in each line of the
delimited file; \code{","} by default (as in a csv file).}

\item{cont}{Logical - are the values continuous (\code{TRUE}, default) or
categories (\code{FALSE})? See Details.}

\item{header}{Logical indicating whether the first line of the file is a
header; default \code{TRUE}.}

\item{rownames}{Default (\code{TRUE}) indicates that the first column of the
file includes names of rows. If a character vector is given here, these
names are used as rownames; if \code{FALSE}, no rownames are used.}

\item{overwrite}{Logical: if a file in \code{dir.out} with the given output
file name already exists, should it be overwritten or not? If NULL, the
user will be prompted for input.}
}
\value{
A list of \code{ff_matrix} objects with the environmental data in
numeric format.
}
\description{
This function reads in environmental data, such as DNA methylation data in
the form of beta values, from a delimited file.
}
\details{
Files containing environmental data, such as methylation data, can be large.
This function facilitates memory-efficient storage and access of such data
using the \code{ff} file format. This function reads in delimited data and
converts it to a list consisting of one or more \code{ff_matrix} objects,
while also exporting the data to a fileset consisting of one \code{.ffData}
file and one \code{.RData} file (an \code{ffarchive}). This process can take
time, but it only needs to be performed once since the output files (i.e.,
the \code{.ffData} + \code{.RData} fileset) can be quickly loaded later on
using the \link{envDataLoad} function. The resulting list of \code{ff_matrix}
objects provides fast and memory-efficient access to the environmental data.
}
\section{Details}{
 If \code{file.out} is not given, the default is NULL and
the output filenames are constructed based on the input filenames. The
suffix '_env' is added to the base name and the \code{.ffData} file is
written to disk. This file contains all the information needed to restore
the list with \code{ff_matrix} objects by calling \link{envDataLoad}
function later on.

If \code{cont} is \code{TRUE} (default), the output data will be a list of
ff matrices containing single-precision values. However, before using this
data as stratification values, the user needs to create categories - this
can be done manually or with the provided \link{envDataCategorize}
function.
}

